+++ src/models/outputgooglecloudobservability.ts
@@ -156,7 +156,7 @@ export type OutputGoogleCloudObservability = {
   keepAliveTime?: number | undefined;
   tls?: TlsSettingsClientSideTypeExtended | undefined;
   /**
-   * Max number of events to include in the request body. Default is 0 (unlimited).
+   * Max number of events to include in the request body. Default is 0 (unlimited). Use to keep outgoing data points within GCO request limits. For metrics, combine with the OTLP Metrics function batchSize.
    */
   maxPayloadEvents?: number | undefined;
   /**