+++ RELEASES.md
@@ -1809,3 +1809,13 @@ Based on:
 - [typescript v0.9.0-alpha.16] .
 ### Releases
 - [NPM v0.9.0-alpha.16] https://www.npmjs.com/package/cribl-control-plane/v/0.9.0-alpha.16 - .
+
+## 2026-06-17 04:11:10
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.778.0 (2.904.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.9.0-alpha.17] .
+### Releases
+- [NPM v0.9.0-alpha.17] https://www.npmjs.com/package/cribl-control-plane/v/0.9.0-alpha.17 - .